Refactor CallCoordinator environment initialization
