+++ GliaWidgets/Sources/Coordinators/Call/CallCoordinator.swift
@@ -114,35 +114,3 @@ private extension CallCoordinator {
         }
     }
 }
-
-extension CallCoordinator {
-    struct Environment {
-        var fetchFile: CoreSdkClient.FetchFile
-        var downloadSecureFile: CoreSdkClient.DownloadSecureFile
-        var uploadFileToEngagement: CoreSdkClient.UploadFileToEngagement
-        var fileManager: FoundationBased.FileManager
-        var data: FoundationBased.Data
-        var date: () -> Date
-        var gcd: GCD
-        var createThumbnailGenerator: () -> QuickLookBased.ThumbnailGenerator
-        var createFileDownload: FileDownloader.CreateFileDownload
-        var fromHistory: () -> Bool
-        var fetchSiteConfigurations: CoreSdkClient.FetchSiteConfigurations
-        var getCurrentEngagement: CoreSdkClient.GetCurrentEngagement
-        var timerProviding: FoundationBased.Timer.Providing
-        var submitSurveyAnswer: CoreSdkClient.SubmitSurveyAnswer
-        var uuid: () -> UUID
-        var uiApplication: UIKitBased.UIApplication
-        var uiScreen: UIKitBased.UIScreen
-        var notificationCenter: FoundationBased.NotificationCenter
-        var fetchChatHistory: CoreSdkClient.FetchChatHistory
-        var createFileUploadListModel: SecureConversations.FileUploadListViewModel.Create
-        var createSendMessagePayload: CoreSdkClient.CreateSendMessagePayload
-        var proximityManager: ProximityManager
-        var log: CoreSdkClient.Logger
-        var snackBar: SnackBar
-        var cameraDeviceManager: CoreSdkClient.GetCameraDeviceManageable
-        var flipCameraButtonStyle: FlipCameraButtonStyle
-        var alertManager: AlertManager
-    }
-}
